move RecordingClient
