+++ crates/iota-sdk-transaction-builder/src/builder/mod.rs
@@ -1717,134 +1717,12 @@ mod tests {
         assert!(matches!(split.amounts[1], Argument::Input(0)));
     }
 
-    /// A [`TestClient`] that records how the builder asked for objects, and can
-    /// report chosen ids as missing.
-    #[cfg(feature = "test-client")]
-    #[derive(Clone, Default)]
-    struct RecordingClient {
-        /// The ids of each `objects_by_id` call, in call order.
-        batches: std::sync::Arc<std::sync::Mutex<Vec<Vec<ObjectId>>>>,
-        /// The ids of each single-object `object` call, in call order.
-        singles: std::sync::Arc<std::sync::Mutex<Vec<ObjectId>>>,
-        /// Ids to report as missing instead of fabricating an object.
-        missing: Vec<ObjectId>,
-    }
-
-    #[cfg(feature = "test-client")]
-    impl RecordingClient {
-        fn batches(&self) -> Vec<Vec<ObjectId>> {
-            self.batches.lock().unwrap().clone()
-        }
-
-        fn singles(&self) -> Vec<ObjectId> {
-            self.singles.lock().unwrap().clone()
-        }
-    }
-
-    #[cfg(feature = "test-client")]
-    impl TransactionBuilderClient for RecordingClient {
-        type Error = crate::TestClientError;
-        type DryRunResult = ();
-
-        async fn object(
-            &self,
-            object_id: ObjectId,
-            version: impl Into<Option<Version>>,
-        ) -> Result<Option<Object>, Self::Error> {
-            self.singles.lock().unwrap().push(object_id);
-            if self.missing.contains(&object_id) {
-                return Ok(None);
-            }
-            crate::TestClient.object(object_id, version).await
-        }
-
-        async fn objects_by_id(
-            &self,
-            object_ids: &[(ObjectId, Option<Version>)],
-        ) -> Result<Vec<Option<Object>>, Self::Error> {
-            self.batches
-                .lock()
-                .unwrap()
-                .push(object_ids.iter().map(|(id, _)| *id).collect());
-            let mut objects = Vec::with_capacity(object_ids.len());
-            for (object_id, _) in object_ids {
-                objects.push(if self.missing.contains(object_id) {
-                    None
-                } else {
-                    crate::TestClient.object(*object_id, None).await?
-                });
-            }
-            Ok(objects)
-        }
-
-        async fn objects(
-            &self,
-            struct_tag: Option<StructTag>,
-            owner: Address,
-            cursor: Option<Vec<u8>>,
-            limit: Option<usize>,
-        ) -> Result<crate::ObjectsPage, Self::Error> {
-            crate::TestClient
-                .objects(struct_tag, owner, cursor, limit)
-                .await
-        }
-
-        async fn transaction(
-            &self,
-            digest: iota_types::TransactionDigest,
-        ) -> Result<Option<iota_types::SignedTransaction>, Self::Error> {
-            crate::TestClient.transaction(digest).await
-        }
-
-        async fn transaction_effects(
-            &self,
-            digest: iota_types::TransactionDigest,
-        ) -> Result<Option<TransactionEffects>, Self::Error> {
-            crate::TestClient.transaction_effects(digest).await
-        }
-
-        async fn reference_gas_price(
-            &self,
-            epoch: impl Into<Option<u64>>,
-        ) -> Result<Option<u64>, Self::Error> {
-            crate::TestClient.reference_gas_price(epoch).await
-        }
-
-        async fn estimate_tx_budget(&self, tx: &Transaction) -> Result<Option<u64>, Self::Error> {
-            crate::TestClient.estimate_tx_budget(tx).await
-        }
-
-        async fn dry_run_tx(
-            &self,
-            tx: &Transaction,
-            skip_checks: bool,
-        ) -> Result<Self::DryRunResult, Self::Error> {
-            crate::TestClient.dry_run_tx(tx, skip_checks).await
-        }
-
-        async fn execute_tx(
-            &self,
-            signatures: &[iota_types::UserSignature],
-            tx: &Transaction,
-            wait_for: impl Into<Option<WaitForTx>>,
-        ) -> Result<TransactionEffects, Self::Error> {
-            crate::TestClient.execute_tx(signatures, tx, wait_for).await
-        }
-
-        async fn wait_for_tx(
-            &self,
-            digest: iota_types::TransactionDigest,
-            wait_for: WaitForTx,
-        ) -> Result<(), Self::Error> {
-            crate::TestClient.wait_for_tx(digest, wait_for).await
-        }
-    }
-
     #[cfg(feature = "test-client")]
     mod input_resolution {
         use iota_types::Version;
 
-        use super::{super::*, RecordingClient};
+        use super::*;
+        use crate::RecordingClient;
 
         fn object_id(seed: u8) -> ObjectId {
             ObjectId::new([seed; ObjectId::LENGTH])
